Removed unused classes and moves utility to the correct folder.
