--- app/src/main/java/com/example/soullinkhelper/interfaces/PokemonLinkInterface.java
@@ -1,6 +1,0 @@
-package com.example.soullinkhelper.interfaces;
-
-public interface PokemonLinkInterface {
-    String getPokemonName();
-    void setPokemonName(String name);
-}
